Add files via upload
intro.txt contains the ASCII visuals for the opening of the game. As for the C file "playerPicks.c," this contains a function to print ASCII text, along with prompting the user if they would like to start. I have yet to add in any time of error-handling, etc, but I wanted to get your guys' opinions first before I went on.